Conflicts:
	src/terrama2/impl/DataAccessorPostGIS.cpp
	src/terrama2/services/analysis/core/AnalysisExecutor.cpp
	src/terrama2/services/analysis/core/AnalysisExecutor.hpp

+++ src/terrama2/services/analysis/core/AnalysisExecutor.hpp
@@ -158,7 +158,7 @@ namespace terrama2
 
           signals:
             //! Signal to notify that a analysis execution has finished.
-            void analysisFinished(int, bool);
+            void analysisFinished(size_t, bool);
 
         };
